// src/app/models/enums.ts
export enum Cargo {
  PROGRAMADOR = 'PROGRAMADOR',
  INGENIERO_SISTEMAS = 'INGENIERO_SISTEMAS',
  ANALISTA = 'ANALISTA',
  DISENADOR_UX_UI = 'DISENADOR_UX_UI',
  QA_TESTER = 'QA_TESTER',
  DEVOPS = 'DEVOPS',
  JEFE_DE_PROYECTO = 'JEFE_DE_PROYECTO'
}

export enum EstadoProyecto {
  PENDIENTE = 'PENDIENTE',
  EN_PROGRESO = 'EN_PROGRESO',
  COMPLETADO = 'COMPLETADO',
  CANCELADO = 'CANCELADO'
}

export enum EstadoRegistro {
  ACTIVO = 'ACTIVO',
  INACTIVO = 'INACTIVO'
}

export enum TipoDocumento {
  DNI = 'DNI',
  CARNET_EXTRANJERIA = 'CARNET_EXTRANJERIA',
  RUC = 'RUC',
  RIF = 'RIF'
}